Global Conftest Refactor
